+++ gql-server/package.json
@@ -16,6 +16,7 @@
     "apollo-server": "^2.0.5",
     "apollo-server-express": "^2.0.4",
     "date-fns": "^2.0.0-alpha.16",
+    "esm": "^3.0.82",
     "faker": "^4.1.0",
     "graphql": "^0.13.2",
     "graphql-redis-subscriptions": "^1.5.0",
